// todo-webapp/src/generated/todo-api.ts
export interface components {
    schemas: {
        Todo: {
            id: string;
            title: string;
            done: boolean;
            /** Format: date-time */
            createdAt: string;
        };
        NewTodo: {
            title: string;
        };
        TodoUpdate: {
            done: boolean;
        };
        TodoList: {
            /** @description total matching items */
            count: number;
            /** @description relative URI of the next page */
            next?: string | null;
            /** @description relative URI of the previous page */
            previous?: string | null;
            data: components["schemas"]["Todo"][];
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}